refactor(providers): rename registry to provider runtime
